+++ inpeak/src/main/java/com/blooming/inpeak/answer/dto/request/AnswerCreateRequest.java
@@ -7,7 +7,7 @@ public record AnswerCreateRequest (
     @NotNull
     String audioFile,
     @NotNull
-    int time,
+    Long time,
     @NotNull
     Long questionId,
     @NotNull
